kafka_indexer/requirements.txt
kafka-python
elasticsearch==8.12.0